story-book-maker/pyproject.toml
[project]
name = "story-book-maker"
version = "0.1.0"
description = "Google ADK를 사용한 어린이 동화책 제작 에이전트"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.22.1",
    "google-genai>=1.59.0",
    "litellm>=1.81.0",
    "openai>=2.15.0",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
]
